+++ powershell/jaxrs/src/test/java/com/redhat/rhevm/api/powershell/model/PowerShellVmTest.java
@@ -41,7 +41,7 @@ public class PowerShellVmTest extends PowerShellModelTest {
         assertEquals(v.getCluster().getId(), clusterId);
     }
 
-    private void testDisk(Disk d, Integer size, DiskType type, DiskStatus status, DiskInterface iface, DiskFormat format, Boolean sparse, Boolean bootable, Boolean wipeAfterDelete, Boolean propagateErrors) {
+    private void testDisk(Disk d, Long size, DiskType type, DiskStatus status, DiskInterface iface, DiskFormat format, Boolean sparse, Boolean bootable, Boolean wipeAfterDelete, Boolean propagateErrors) {
         assertEquals(d.getSize(), size);
         assertEquals(d.getType(), type);
         assertEquals(d.getStatus(), status);
@@ -94,7 +94,7 @@ public class PowerShellVmTest extends PowerShellModelTest {
         assertNotNull(vm.getDevices());
         assertEquals(vm.getDevices().getDisks().size(), 1);
 
-        testDisk(vm.getDevices().getDisks().get(0), 683622400, DiskType.SYSTEM, DiskStatus.OK, DiskInterface.IDE, DiskFormat.RAW, true, true, null, null);
+        testDisk(vm.getDevices().getDisks().get(0), 683622400L, DiskType.SYSTEM, DiskStatus.OK, DiskInterface.IDE, DiskFormat.RAW, true, true, null, null);
 
         data = readFileContents("interfaces.data");
         assertNotNull(data);
